Changes before error encountered
Agent-Logs-Url: https://github.com/sb224sc-HT22-VT27/4dt907/sessions/5fcc34b5-2f2d-4b2b-a902-132825c3d30e

+++ src/backend/app/schemas/squat.py
@@ -31,7 +31,7 @@ class FrameAnalysisResult(BaseModel):
     start_stop: int
     predicted_z: Dict[str, float]
     good_bad_score: Optional[float] = None
-    squat_score: Optional[float] = None
+    squat_score: Optional[int] = None
 
 
 class SessionAnalysisResponse(BaseModel):
